Fix the inference problem.
The strategy of statically constructing models in training or
inference mode turned out to not really be practical. The resulting
models don't map one-to-one to each-other, so it's really hard to
load weights from a training-mode model into an inference-mode one.
That kind of defeats the purpose of this whole exercise.

I have moved to a system that allows for dynamic selection of training
vs. inference mode in a single unified model architecture. This is kind
of hacky in some ways and annoying to use, but at least it works.